UID is now obtained from admin
My plan is to use this to verify if user is admin, but also I might be able to use this tutorial on role based access control to hide certain elements from non-admin users. I'm not exactly sure best way to implement this until we have the events tab fully completed.